notification_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
psycopg2==2.9.9
requests==2.31.0

pytz==2023.3.post1
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

drf-spectacular==0.26.5

pytest==7.4.3
pytest-django==4.7.0
pytest-factoryboy==2.6.0